listings ch09
