+++ listings/ch09/listing_11.js
@@ -1,22 +1,5 @@
-export function mountDOM(vdom, parentEl, index/*--add--*/, hostComponent = null/*--add--*/) {
-  switch (vdom.type) {
-    case DOM_TYPES.TEXT: {
-      createTextNode(vdom, parentEl, index)
-      break
-    }
-
-    case DOM_TYPES.ELEMENT: {
-      createElementNode(vdom, parentEl, index/*--add--*/, hostComponent/*--add--*/)
-      break
-    }
-
-    case DOM_TYPES.FRAGMENT: {
-      createFragmentNodes(vdom, parentEl/*--add--*/, hostComponent/*--add--*/)
-      break
-    }
-
-    default: {
-      throw new Error(`Can't mount DOM of type: ${vdom.type}`)
-    }
-  }
+// Safely evaluate `hasOwnProperty` calls.
+// https://eslint.org/docs/latest/rules/no-prototype-builtins
+export function hasOwnProperty(obj, prop) {
+  return Object.prototype.hasOwnProperty.call(obj, prop)
 }
