+++ listings/ch09/listing_13.js
@@ -1,6 +1,43 @@
-function addProps(el, props, vdom/*--add--*/, hostComponent/*--add--*/) {
-  const { on: events, ...attrs } = props
+// --add--
+import { toPromise } from './utils/promises'
+// --add--
 
-  vdom.listeners = addEventListeners(events, el/*--add--*/, hostComponent/*--add--*/)
-  setAttributes(el, attrs)
+export function addEventListeners(
+  listeners = {},
+  el,
+  /*--add--*/hostComponent = null/*--add--*/
+) {
+  const addedListeners = {}
+
+  Object.entries(listeners).forEach(([eventName, handler]) => {
+    const listener = addEventListener(eventName, handler, el/*--add--*/, hostComponent/*--add--*/)
+    addedListeners[eventName] = listener
+  })
+
+  return addedListeners
+}
+
+export function addEventListener(
+  eventName,
+  handler,
+  el,
+  /*--add--*/hostComponent = null/*--add--*/
+) {
+  // --remove--
+  el.addEventListener(eventName, handler)
+  return handler
+  // --remove--
+  // --add--
+  async function asyncHandler() {
+    await toPromise(
+      hostComponent
+        ? handler.call(hostComponent, ...arguments)
+        : handler(...arguments)
+    )
+  }
+
+  el.addEventListener(eventName, asyncHandler)
+
+  return asyncHandler
+  // --add--
 }
