+++ listings/ch09/listing_14.js
@@ -1,8 +1,3 @@
-function createFragmentNodes(vdom, parentEl/*--add--*/, hostComponent/*--add--*/) {
-  const { children } = vdom
-  vdom.el = parentEl
-
-  children.forEach((child) =>
-    mountDOM(child, parentEl, null/*--add--*/, hostComponent/*--add--*/)
-  )
+export function toPromise(value) {
+  return Promise.resolve(value)
 }
