+++ listings/ch09/listing_15.js
@@ -1,27 +1,22 @@
-export function patchDOM(oldVdom, newVdom, parentEl/*--add--*/, hostComponent = null/*--add--*/) {
-  if (!areNodesEqual(oldVdom, newVdom)) {
-    const index = Array.from(parentEl.childNodes).indexOf(oldVdom.el)
-    destroyDOM(oldVdom)
-    mountDOM(newVdom, parentEl, index/*--add--*/, hostComponent/*--add--*/)
-
-    return newVdom
-  }
-
-  newVdom.el = oldVdom.el
-
-  switch (newVdom.type) {
+export function mountDOM(vdom, parentEl, index/*--add--*/, hostComponent = null/*--add--*/) {
+  switch (vdom.type) {
     case DOM_TYPES.TEXT: {
-      patchText(oldVdom, newVdom)
-      return newVdom
+      createTextNode(vdom, parentEl, index)
+      break
     }
 
     case DOM_TYPES.ELEMENT: {
-      patchElement(oldVdom, newVdom/*--add--*/, hostComponent/*--add--*/)
+      createElementNode(vdom, parentEl, index/*--add--*/, hostComponent/*--add--*/)
       break
     }
+
+    case DOM_TYPES.FRAGMENT: {
+      createFragmentNodes(vdom, parentEl/*--add--*/, hostComponent/*--add--*/)
+      break
+    }
+
+    default: {
+      throw new Error(`Can't mount DOM of type: ${vdom.type}`)
+    }
   }
-
-  patchChildren(oldVdom, newVdom/*--add--*/, hostComponent/*--add--*/)
-
-  return newVdom
 }
