+++ listings/ch09/listing_17.js
@@ -1,30 +1,6 @@
-function patchEvents(
-  el,
-  oldListeners = {},
-  oldEvents = {},
-  newEvents = {},
-  // --add--
-  hostComponent
-  // --add--
-) {
-  const { removed, added, updated } = objectsDiff(oldEvents, newEvents)
+function addProps(el, props, vdom/*--add--*/, hostComponent/*--add--*/) {
+  const { on: events, ...attrs } = props
 
-  for (const eventName of removed.concat(updated)) {
-    el.removeEventListener(eventName, oldListeners[eventName])
-  }
-  const addedListeners = {}
-
-  for (const eventName of added.concat(updated)) {
-    const listener = addEventListener(
-      eventName,
-      newEvents[eventName],
-      el,
-      // --add--
-      hostComponent
-      // --add--
-    )
-    addedListeners[eventName] = listener
-  }
-
-  return addedListeners
+  vdom.listeners = addEventListeners(events, el/*--add--*/, hostComponent/*--add--*/)
+  setAttributes(el, attrs)
 }
