+++ listings/ch09/listing_18.js
@@ -1,56 +1,8 @@
-function patchChildren(oldVdom, newVdom/*--add--*/, hostComponent/*--add--*/) {
-  const oldChildren = extractChildren(oldVdom)
-  const newChildren = extractChildren(newVdom)
-  const parentEl = oldVdom.el
+function createFragmentNodes(vdom, parentEl/*--add--*/, hostComponent/*--add--*/) {
+  const { children } = vdom
+  vdom.el = parentEl
 
-  const diffSeq = arraysDiffSequence(
-    oldChildren,
-    newChildren,
-    areNodesEqual
+  children.forEach((child) =>
+    mountDOM(child, parentEl, null/*--add--*/, hostComponent/*--add--*/)
   )
-  
-  for (const operation of diffSeq) {
-    const { from, index, item } = operation
-
-    switch (operation.op) {
-      case ARRAY_DIFF_OP.ADD: {
-        mountDOM(item, parentEl, index/*--add--*/, hostComponent/*--add--*/)
-        break
-      }
-
-      case ARRAY_DIFF_OP.REMOVE: {
-        destroyDOM(item)
-        break
-      }
-        
-      case ARRAY_DIFF_OP.MOVE: {
-        const el = oldChildren[from].el
-        const elAtTargetIndex = parentEl.childNodes[index]
-
-        parentEl.insertBefore(el, elAtTargetIndex)
-        patchDOM(
-          oldChildren[from],
-          newChildren[index],
-          parentEl,
-          // --add--
-          hostComponent
-          // --add--
-        )
-
-        break
-      }
-
-      case ARRAY_DIFF_OP.NOOP: {
-        patchDOM(
-          oldChildren[from],
-          newChildren[index],
-          parentEl,
-          // --add--
-          hostComponent
-          // --add--
-        )
-        break
-      }
-    }
-  }
 }
